fameOn/requirements.txt
django-allauth
django 1.9.5
